create reusable FakeSyncStatusProvider for tests
